refactor: pass portfolio scroll gettering logic to header
